Ensure CMake finds CUDA libraries from host packages
Consider a conda recipe that looks like:
```
requirements:
  build:
    - {{ compiler('c') }}
    - {{ compiler('cxx') }}
    - {{ compiler('cuda') }}
    - cmake >=3.26.4
    - sysroot_{{ target_platform }} 2.17
    - cuda-version =12.0
  host:
    - libcublas-dev
    - libcublas-static
    - cuda-version =12.0
```

We have placed the CUDA compiler in `build` as the compiler supports
cross-compiling. We place the cublas libraries in `host` since
it is only for the target platform we are building. Now since
the compiler is in `build` it isn't aware of the `host` libraries
and doesn't report those locations as implicit search dirs.

So for CMake to find the cublas libraries we need to extend
CMAKE_FIND_ROOT_PATH to search `$PREFIX/${targetsDir}`.